SVCPLAN-7606: update dcgmi model detection to handle H200 GPUs
lspci does not show the model for H200 GPUs. Thus we'll switch to a new
detection method that uses nvidia-smi to get the model name. It actually
stores this in a file so it only has to run nvidia-smi once and thereafter
checks the file. This involves adjusting the case statements to use
the more user-friendly model names returned by nvidia-smi rather than
the "under-the-hood" model names returned by lscpi for other models.