SVCPLAN-8288: BREAKING - support alternate package name
BREAKING: This update changes the default version of DCGM
from v3 to v4 (with CUDA 12 support). Please see README.md
for instructions for upgrading from v3 and/or need to support
different/additional versions of CUDA.

This update adjusts a Puppet 'require' dependency to use use the DCGM
package variable instead of hard-coding the name, to support customizing
which packages are installed while still enforcing the dependency.